+++ CHANGELOG.md
@@ -56,7 +56,7 @@ This is the rolling changelog for TShock for Terraria. Use past tense when addin
 * Added `TSPlayer` to `GetDataHandlers.GemLockToggle`. (@hakusaro)
 * Added `GetDataHandlers.PlaceItemFrame` hook and related arguments. (@hakusaro)
 * Added `TSPlayer.IsBouncerThrottled()`. (@hakusaro)
-* Added `TSPlayer.CheckIgnores()` and removed `TShock.CheckIgnores(TSPlayer)`. (@hakusaro)
+* Added `TSPlayer.IsBeingDisabled()` and removed `TShock.CheckIgnores(TSPlayer)`. (@hakusaro)
 
 ## TShock 4.3.25
 * Fixed a critical exploit in the Terraria protocol that could cause massive unpreventable world corruption as well as a number of other problems. Thanks to @bartico6 for reporting. Fixed by the efforts of @QuiCM, @hakusaro, and tips in the right directioon from @bartico6.